gedcom_processor/pyproject.toml
[project]
name = "gedcom-processor"
version = "0.1.0"
description = "Generate TRMNL family JSON from GEDCOM files"
requires-python = ">=3.11"
dependencies = [
    "ged4py>=0.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
generate = "generate:main"
